Make grid aggregation consider map as pixels
This commit finish WIP from previous commit. Applying this will
change how grid aggregation is done. Maps are instead considered as
a set of pixels with area that can overlap with the extents of the
grid cells. This changes the interpretation of the resulting maps,
but only slightly. This should be more robust, but is perhaps less
faithful to how xtgeo.RegularSurface represents data.

Missing:
- Rename "node" to "pixel" where appropriate, where pixel refers to
  the area covered by a pixel, and not the point location of a node

- Properly handle xtgeo.RegularSurface as the map template input. In
  the current implementation, the template will be interpreted as
  pixels, meaning that the output will be smaller (-1) in both x and
  y directions. How do we handle the corners, though?

- Misc. TODOs for cleaner code